// packages/graphback-codegen-schema/src/writer/schemaFormatters.ts
import { SchemaFormatter } from '../SchemaFormatter'

const noteString = '// NOTE: This schema was generated by Graphback and should not be changed manually'

/**
 * Typescript string template that formats schema into common js module that can be imported 
 * easily in server side application
 */
export const tsSchemaFormatter: SchemaFormatter = {
  format: (schemaString: string) => {
    return `
${noteString}
export const schemaString = \`
${schemaString}
\`;
`
  }
}

/**
 * JS string template that formats schema into common js module that can be imported 
 * easily in server side application
 */
export const jsSchemaFormatter: SchemaFormatter = {
  format: (schemaString: string) => {
    return `
${noteString}
const schemaString = \`
${schemaString}
\`;

module.exports = { schemaString };
`
  }
}


/**
 * GQL string template that returns schema in original form
 */
export const gqlSchemaFormatter: SchemaFormatter = {
  format: (schemaString: string) => {
    return `${noteString}\n${schemaString}`;
  }
}